implement SlotManager; devices table; DeviceMap and SystemConfig classes; initialize system devices by iterating through devs in SystemConfig; OSD uses SlotManager to get slot device information

+++ src/devices/prodos_clock/prodos_clock.hpp
@@ -28,4 +28,4 @@ struct prodos_clock_state {
     char buf[64];    
 };
 
-void init_slot_prodosclock(cpu_state *cpu, int slot);
+void init_slot_prodosclock(cpu_state *cpu, uint8_t slot);
